refactor: relocated card's testing data into tests directory
